feat: make yprofile multi command style
The first step to make yprofile to become the multi command style

BREAKING CHANGE: yprofile now change `yprofile INPUT_FILE PROFILE` to `yprofile generate|gen
INPUT_FILE PROFILE`
