+++ bin/run
@@ -1,13 +1,5 @@
 #!/usr/bin/env node
 
-const fs = require('fs')
-const path = require('path')
-const project = path.join(__dirname, '../tsconfig.json')
-const dev = fs.existsSync(project)
-
-if (dev) {
-  require('ts-node').register({project})
-}
-
-require(`../${dev ? 'src' : 'lib'}`).run()
+require('@oclif/command').run()
+.then(require('@oclif/command/flush'))
 .catch(require('@oclif/errors/handle'))
