+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@holytiny/yprofile",
-  "description": "A tool to generate yaml file use 'in-block' profile.",
+  "description": "A tool to generate yaml file use in-block profile.",
   "version": "1.0.0",
   "author": "sake",
   "bin": {
@@ -51,7 +51,11 @@
   "license": "MIT",
   "main": "lib/index.js",
   "oclif": {
-    "bin": "yprofile"
+    "commands": "./lib/commands",
+    "bin": "yprofile",
+    "plugins": [
+      "@oclif/plugin-help"
+    ]
   },
   "repository": {
     "type": "git",
